subtree: fix argument handling in check_parents
check_parents was taking all of its arguments as a single string,
and erroneously passing them to cache_miss as a single string. For
commits with a single parent this would succeed, but whenever a merge
commit was processed, cache_miss would be passed "parent1 parent2"
instead of "parent1" "parent2" and fail, leading to unecessary rechecks
of the parent commits.

For consistency, take multiple arguments in check_parents,
and pass all of them to cache_miss separately.

Signed-off-by: James Limbouris <james@digitalmatter.com>