sparse-checkout: bash custom tab completion tests
Add tests for custom tab completion of the sparse-checkout command in
bash. This command currently relies on the tab completion provided by the
CLI. This is not ideal because:

1. git sparse-checkout <TAB> results in an incomplete list of subcommands.
2. git sparse-checkout --<TAB> does not populate the help option.
3. Options for subcommands are not populated.
4. git sparse-checkout set <TAB> and git sparse-checkout add <TAB> show
both file names and directory names.

This change adds tests to highlight these issues. These tests currently
fail but will succeed with the sparse-checkout additions to
git-completion.bash in the next commit in this series.

Signed-off-by: Lessley Dennington <lessleydennington@gmail.com>